fix: correct connection between UI and graphVM
The main issue is, that graph representation updates only after some strange interaction with UI. I've tried mutliple ways to overcome it. The main - is MutableStateListOf, where I try to store viewmodels of vertices, but it seems to be a bad practice, bc of many other instructions, are needed, to help GraphView connect with it. The last thing about it - if I store a mutableStateList - I need to to modify an element of this list, to call the recomposition -> adding a new element won't do it))). In the end of all - I add mutableState 'updateIsRequired' that is used as recomposition indicator (see next commit).